requirements.txt
Flask==2.2.5
pymongo==4.4.1
openai>=0.27.0
python-dotenv==1.0.0
